requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
joblib==1.4.2
numpy==2.0.2
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
starlette==0.45.3